+++ assets/src/turbo_stream_controller.ts
@@ -16,24 +16,34 @@ import { connectStreamSource, disconnectStreamSource } from '@hotwired/turbo';
 export default class extends Controller {
     static values = {
         topic: String,
+        topics: Array,
         hub: String,
     };
     es: EventSource | undefined;
     url: string | undefined;
 
     declare readonly topicValue: string;
+    declare readonly topicsValue: string[];
     declare readonly hubValue: string;
     declare readonly hasHubValue: boolean;
     declare readonly hasTopicValue: boolean;
+    declare readonly hasTopicsValue: boolean;
 
     initialize() {
         const errorMessages: string[] = [];
         if (!this.hasHubValue) errorMessages.push('A "hub" value pointing to the Mercure hub must be provided.');
-        if (!this.hasTopicValue) errorMessages.push('A "topic" value must be provided.');
+        if (!this.hasTopicValue && !this.hasTopicsValue)
+            errorMessages.push('Either "topic" or "topics" value must be provided.');
         if (errorMessages.length) throw new Error(errorMessages.join(' '));
 
         const u = new URL(this.hubValue);
-        u.searchParams.append('topic', this.topicValue);
+        if (this.hasTopicValue) {
+            u.searchParams.append('topic', this.topicValue);
+        } else {
+            this.topicsValue.forEach((topic) => {
+                u.searchParams.append('topic', topic);
+            });
+        }
 
         this.url = u.toString();
     }